205201 To map the business rules for unit of sale type as Unit and process S-100 v1.1 events only (#306)
* Added code for S100 Create rules AB#205201

* Reverted contentType rule AB#205201

* AB#205201-Added test cases for folio

* AB#205201-Added file changes

* Updated V1 to V1.1 in Unit test json file AB#205201

* Rename file AB#205201

* AB# 205201 Added Automation Test Changes

* AB# 205201 Added minor changes

* AB#205201 Added minor Change

* AB# 205201 Added Test Changes

* AB# 205201 Added minor changes

* AB# 205201 Fixed Review Comment

* Fixed minor review comment AB#205201

---------

+++ src/UKHO.ERPFacade.Common/Models/CloudEvents/S100Event/S100BoundingBox.cs
@@ -9,9 +9,9 @@ namespace UKHO.ERPFacade.Common.Models.CloudEvents.S100Event
         [JsonProperty("northLimit")]
         public double NorthLimit { get; set; }
         [JsonProperty("southLimit")]
-        public int SouthLimit { get; set; }
+        public double SouthLimit { get; set; }
         [JsonProperty("eastLimit")]
-        public int EastLimit { get; set; }
+        public double EastLimit { get; set; }
         [JsonProperty("westLimit")]
         public double WestLimit { get; set; }
     }
